dimash

+++ admin/js/index.js
@@ -1,17 +1,5 @@
-const createForm = document.querySelector("#create-form");
-const loginButtons = document.querySelectorAll(".login")
-const exit = document.querySelector(".signOut")
-const statusEmail = document.querySelector(".status-email")
-const statusFname = document.querySelector(".status-fname")
-const status = document.querySelector(".status")
-const statusGuest = document.querySelector(".status")
-const calculator = document.querySelector(".calculator")
-const loginInputs = document.getElementById("login-inputs")
-const userOrders = document.querySelector(".user_orders")
+const createForm = document.querySelector("#create-form"), loginButtons = document.querySelectorAll(".login"), exit = document.querySelector(".signOut"),statusEmail = document.querySelector(".status-email"),statusFname = document.querySelector(".status-fname"),status = document.querySelector(".status"),statusGuest = document.querySelector(".status"),calculator = document.querySelector(".calculator"),loginInputs = document.getElementById("login-inputs"),userOrders = document.querySelector(".user_orders")
 
-function gener() {
-    showProducts();
-}
 
 function createAdmin(cancelAd) {
     let makeAdminForm = document.getElementById("make-admin-form")
@@ -150,7 +138,7 @@ function addProduct() {
     db.collection("products").add({
         header: createProductForm.children[0].value,
         text: createProductForm.children[1].value,
-    }).then(function(docRef) {
+    }).then(function (docRef) {
         createProductForm.children[4].innerHTML = "Товар добавлен"
         file = createProductForm.children[2].files[0];
         storageRef = firebase.storage().ref(`${file.name}`);
@@ -161,37 +149,61 @@ function addProduct() {
                 }).catch(() => {
                     console.log("Изображение не добавлено")
                 })
-        })
+            })
         })
     }
     )
 }
 
-function showProducts() {
-    db.collection("products").get().then(snapshot => {
-        try {
-            let products = document.querySelector(".product-list")
-            snapshot.docs.forEach(function callback(doc, index) {
-                const html = `
-            <div style="border: 5px solid;" class="product${index}">
-            <p>${doc.data().header}</p>
-            <p>${doc.data().text}</p>
-            <img src="${doc.data().image}">
-            <button id="btn${index}">Удалить</button>
-            </div>
-            `
-
-            products.innerHTML += html
-                    document.getElementById("btn" + index).addEventListener('click', (e) => {
-                        console.log(doc.id)
-                        db.collection("products").doc(doc.id).delete().then(() => {
-                            console.log("Товар удален");
-                        })
+function showProducts(data) {
+    let products = document.querySelector(".product-list")
+    try {
+        let html = '';
+        data.forEach(function callback(doc, index) {
+            const list = `<div style="border: 5px solid;" class="product${index}">
+                <input value="${doc.data().header}">
+                <input value="${doc.data().text}">
+                <img style="width:100px; height:100px;" src="${doc.data().image}">
+                <input type="file">
+                <button id="btnd${index}">Удалить</button>
+                <button id="btnu${index}">Редактировать</button>
+                </div>`
+                html += list
+                firebase.auth().onAuthStateChanged((user) => {
+                    if (user) {
+                        document.querySelector(".product" + index).addEventListener('click', (e) => {
+                            if (e.target.id == "btnd" + index) {
+                                db.collection("products").doc(doc.id).delete().then(() => {
+                                    console.log("Товар удален")
+                                })
+                            }
+                            if (e.target.id == "btnu" + index) {
+                                file = document.querySelector(".product" + index).children[3].files[0];
+                                storageRef = firebase.storage().ref(`${file.name}`);
+                                storageRef.put(file).then(() => {
+                                    storageRef.getDownloadURL().then((url) => {
+                                        db.collection("products").doc(doc.id).update({
+                                            image: url
+                                        }).catch(() => {
+                                            console.log("Изображение не добавлено")
+                                        })
+                                    })
+                                })
+                            db.collection("products").doc(doc.id).update({
+                                header: document.querySelector(".product" + index).children[0].value,
+                                text: document.querySelector(".product" + index).children[1].value,
+                            }).then(() => {
+                                console.log("Товар обновлен")
+                            })
+                        }
                     })
+                }
             })
-        } catch (e) {
-            console.log(e)
-        }
-    })
+        })
+        products.innerHTML = html
+
+    } catch (e) {
+        console.log(e)
+    }
 }
 //                                  END ADD PRODUCTS
